requirements.txt
#--- FastAPI Core & Extensions ---
fastapi[all]>=0.100.0
pydantic-settings>=2.0.0

#--- Database, Auth & Storage (Official Unified Client) ---
supabase>=2.15.0

#--- Mapping & Geocoding ---
opencage>=2.0.0
timezonefinder>=6.0.0

#--- Calendar Event Generation ---
ics>=0.7
pytz>=2023.3

#--- Security & Utilities ---
# Note: AuthApiError and JWT handling are now natively managed via the supabase client
python-dotenv>=1.0.0
gunicorn>=20.1.0,<22.0.0